fwk: Move signal infrastructure into separate file
This patch cleans up the signalling infrastructure in preparation
for adding the signals API to the multi-threaded framework.
