+++ module/dvfs/src/mod_dvfs.c
@@ -18,6 +18,7 @@
 #include <fwk_mm.h>
 #include <fwk_module.h>
 #include <fwk_module_idx.h>
+#include <fwk_signal.h>
 #include <fwk_status.h>
 #include <fwk_thread.h>
